pyproject.toml
[project]
name = "docsmith"
version = "1.1.0"
description = "YAML-to-Word document generator for automated pipelines and LLM-driven content creation"
authors = [{name = "Development Team"}]
readme = "README.md"
requires-python = ">=3.13"
license = "GPL-3.0-or-later"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business",
    "Topic :: Text Processing :: Markup",
    "Typing :: Typed",
]

dependencies = [
    "python-docx",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pillow",
    "pre-commit",
    "pytest",
    "ruff",
]

[project.scripts]
docsmith = "docsmith.cli:main"

[project.urls]
Homepage = "https://github.com/dawsonlp/docsmith"
Repository = "https://github.com/dawsonlp/docsmith"
Issues = "https://github.com/dawsonlp/docsmith/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/docsmith"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
src = ["src"]
target-version = "py313"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["docsmith"]